Update README.md
A linear regression analysis was performed for an ecommerce company to help them decide between focusing on their app experience or their website. The model showed that the app currently has greater profitability for one more minute increase compared to the time on the website. An economical analysis would need to be done to further analyze which of the two would be more profitable: bringing the website up or developing the app further in efforts to increase time on either. Alongside this,however, Length of Membership was the greatest impacter in the amount a customer spent yearly, meaning that the longer the customers stay with the company, the more money the company will make in the long run. Length of Membership should be included in the economic consideration of what to focus on.